исправили задачу 3

+++ zadacha_3/Program.cs
@@ -4,7 +4,7 @@
 // A (7,-5); B (1,-1) -> 7,21
 
 
-double EnterPointCoordinate(string str)
+ int EnterPointCoordinate(string str)
 {
     Console.WriteLine(str);
     double number1 = double.Parse(Console.ReadLine());
